requirements.txt
h5py
numpy
pyyaml
pathlib
tensorflow==2.4.1
tensorflow-gpu==2.4.1
keras==2.3.1
pandas
tqdm
scikit-learn
matplotlib
seaborn
scipy